feat(blog): build system architecture page with reusable ProgressiveTerminal
- Add system architecture page with C1/C2/C3 diagram sections and terminal-voice copy
- Extract ProgressiveTerminal to shared component accepting lines[] and isPaused props
- Create LandingTerminal wrapper to preserve homepage-specific context callbacks
- Move FooterSection to shared components/Footer for reuse across pages
- Remove dead TerminalEvent enum and unused onTerminalEvent* context callbacks
