--- apps/blog/src/app/(home)/LandingSection/TerminalEvent.tsx
@@ -1,12 +1,0 @@
-export enum TerminalEvent {
-  WELCOMING_USER = 'WELCOMING_USER',
-  INITIALIZING_WEBSITE_TITLE = 'INITIALIZING_WEBSITE_TITLE',
-  RENDERING_CONTROL_DEVICES = 'RENDERING_CONTROL_DEVICES',
-  GENERATING_NON_ESSENTIAL_VISUAL_LAYER = 'GENERATING_NON_ESSENTIAL_VISUAL_LAYER',
-  AWAITING_USER_ACTION = 'AWAITING_USER_ACTION',
-  IDLE = 'IDLE',
-  TIME_ELAPSED_EXCEEDS_EXPECTATION = 'TIME_ELAPSED_EXCEEDS_EXPECTATION',
-  PROCESS_DELAYED = 'PROCESS_DELAYED',
-  CLICK_BUTTON_TO_CONTINUE = 'CLICK_BUTTON_TO_CONTINUE',
-  SCROLL_DOWN_PAGE = 'SCROLL_DOWN_PAGE',
-}
